make SecureKeyBackup internal
